Split YAML parsing and validation to separate classes
