+++ src/main/java/org/verifyica/pipeliner/Converter.java
@@ -31,9 +31,9 @@ import org.verifyica.pipeliner.common.ValidatorException;
 import org.verifyica.pipeliner.common.Version;
 import org.verifyica.pipeliner.core.Job;
 import org.verifyica.pipeliner.core.Pipeline;
-import org.verifyica.pipeliner.core.PipelineFactory;
 import org.verifyica.pipeliner.core.Run;
 import org.verifyica.pipeliner.core.Step;
+import org.verifyica.pipeliner.core.parser.PipelineParser;
 import org.yaml.snakeyaml.error.MarkedYAMLException;
 import picocli.CommandLine;
 
@@ -98,7 +98,7 @@ public class Converter implements Runnable {
         String pipelinerWorkingDirectory = file.getAbsoluteFile().getParent();
 
         Console console = new Console();
-        Pipeline pipeline = new PipelineFactory(console).createPipeline(file.getAbsolutePath());
+        Pipeline pipeline = new PipelineParser(console).parse(file.getAbsolutePath());
 
         System.out.println("# pipeline name=[" + pipeline.getName() + "]");
