+++ src/main/java/org/verifyica/pipeliner/core/parser/Parser.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2024-present Pipeliner project authors and contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.verifyica.pipeliner.core.parser;
+
+import org.verifyica.pipeliner.common.Console;
+import org.verifyica.pipeliner.common.Converter;
+import org.verifyica.pipeliner.common.Validator;
+
+/** Class to implement Parser */
+public abstract class Parser {
+
+    /** Console */
+    protected final Console console;
+
+    /** Validator */
+    protected final Validator validator;
+
+    /** Converter */
+    protected final Converter converter;
+
+    /**
+     * Constructor
+     *
+     * @param console console
+     */
+    protected Parser(Console console) {
+        this.console = console;
+        this.validator = new Validator();
+        this.converter = new Converter();
+    }
+}
